[CMake] Improved Lua detection - #808
The default Lua version on many systems is often newer than 5.1
However, such versions seem to be incompatible with OLX.
So search first for "lua5.1" and if it's not found,
search for "lua-5.1" - this works on Fedora, for example.
If neither is found, search for "lua" and finally use default "lua"
but give warnings and suggestions too because errors may be likely.
TODO: Header search path should be configured dynamically too.